requirements.txt
# conda install pytorch==2.2.2 torchvision==0.17.2 torchaudio==2.2.2 pytorch-cuda=11.8 -c pytorch -c nvidia
accelerate==0.28.0
einops==0.8.1
ftfy==6.3.1
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.8.1.78
packaging==25.0
Pillow==11.1.0
pycocotools==2.0.7
pytorch_lightning==2.5.1.post0
regex==2024.11.6
scipy==1.15.2
tensorboard==2.19.0
timm==1.0.15
tqdm==4.67.1
# lmdb==1.7.3